requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
django-environ>=0.11
psycopg2-binary>=2.9
python-decouple>=3.8
Pillow>=10.0
numpy>=1.24
scipy>=1.10
drf-spectacular>=0.27.0
python-dateutil>=2.8

user-agents>=2.2.0

# Production dependencies
gunicorn>=21.2.0
whitenoise>=6.6.0
dj-database-url>=2.1.0
